+++ Aion/DAL/Entities/sp_GetRecruitmentDetailDXNS_Result.cs
@@ -22,7 +22,7 @@ namespace Aion.DAL.Entities
         public decimal ContractBase { get; set; }
         public decimal CurrentContract { get; set; }
         public Nullable<decimal> AverageContract { get; set; }
-        public int OpenVacancies { get; set; }
+        public decimal OpenVacancies { get; set; }
         public Nullable<decimal> RateOfPay { get; set; }
         public Nullable<decimal> TotalBase { get; set; }
         public Nullable<decimal> CurrentBase { get; set; }
